refactor(#2158): split userspace manager.go into core + capabilities + controllers
manager.go had grown to 2186 LOC, over the ~2000-LOC modularity
threshold (docs/engineering-style.md). Per the converged #2158 plan
(docs/research/2158-file-split/plan.md §5.6) this is a pure same-package
code-motion split: move two cohesive groups out of manager.go into new
files in the same `package userspace`.

  - capabilities.go (469 LOC): the config->capability derivation cluster
    (deriveUserspaceConfig, deriveUserspaceCapabilities, the
    userspace*Supported / expandUserspace* / normalizeUserspace* /
    resolveUserspace* helpers). All free functions, no *Manager receiver.
  - controllers.go (144 LOC): the dataplane adapter types
    (userspaceLinkController, userspaceHAOps / managerHAOps,
    userspaceHAController, userspaceSessionStore) and their methods.

manager.go drops to 1594 LOC (Manager struct, lifecycle, Compile,
apply/overlay plumbing, neighbor snapshot, accessors). init() and Boot()
stay in manager.go (no init-order change).

Pure code-motion: functions, methods, and types moved verbatim; the only
non-move lines are the per-file import blocks and package declarations.
No logic change, no exported-API change, no receiver change.

Validation:
  - go build ./pkg/dataplane/userspace/ clean
  - go doc -all ./pkg/dataplane/userspace/ byte-identical before/after
  - gofmt -l clean on all three files
  - full package test suite unchanged (see follow-up audit commit)
